+++ contrib/lips4/gdevl4r.c
@@ -56,9 +56,9 @@ static dev_proc_image_out(lips4_image_out);
           (int)((long)(DEFAULT_WIDTH_10THS) * (xdpi) / 10),\
           (int)((long)(DEFAULT_HEIGHT_10THS) * (ydpi) / 10),\
           xdpi, ydpi, color_bits,\
-          -(lm) * (xdpi), -(tm) * (ydpi),\
-          (lm) * 72.0, (bm) * 72.0,\
-          (rm) * 72.0, (tm) * 72.0\
+          (float)(-(lm) * (xdpi)), (float)(-(tm) * (ydpi)),\
+          (float)((lm) * 72.0), (float)((bm) * 72.0f),\
+          (float)((rm) * 72.0), (float)((tm) * 72.0f)\
         ),\
        lp_device_body_rest_(print_page_copies, image_out),\
            cassetFeed, username, LIPS_PJL_DEFAULT,\
@@ -71,9 +71,9 @@ static dev_proc_image_out(lips4_image_out);
           (int)((long)(DEFAULT_WIDTH_10THS) * (xdpi) / 10),\
           (int)((long)(DEFAULT_HEIGHT_10THS) * (ydpi) / 10),\
           xdpi, ydpi, color_bits,\
-          -(lm) * (xdpi), -(tm) * (ydpi),\
-          (lm) * 72.0, (bm) * 72.0,\
-          (rm) * 72.0, (tm) * 72.0\
+          (float)(-(lm) * (xdpi)), (float)(-(tm) * (ydpi)),\
+          (float)((lm) * 72.0), (float)((bm) * 72.0),\
+          (float)((rm) * 72.0), (float)((tm) * 72.0)\
         ),\
        lp_duplex_device_body_rest_(print_page_copies, image_out),\
   cassetFeed,\
@@ -214,10 +214,10 @@ lips4_open(gx_device * pdev)
 static int
 lips_open(gx_device * pdev, lips_printer_type ptype)
 {
-    int width = pdev->MediaSize[0];
-    int height = pdev->MediaSize[1];
-    int xdpi = pdev->x_pixels_per_inch;
-    int ydpi = pdev->y_pixels_per_inch;
+    int width = (int)pdev->MediaSize[0];
+    int height = (int)pdev->MediaSize[1];
+    int xdpi = (int)pdev->x_pixels_per_inch;
+    int ydpi = (int)pdev->y_pixels_per_inch;
 
     /* Paper Size Check */
     if (width <= height) {	/* portrait */
@@ -649,8 +649,8 @@ lips4type_print_page_copies(gx_device_printer * pdev, gp_file * prn_stream, int 
           return_error(gs_error_VMerror);
 
         if (lprn->NegativePrint) {
-          int rm = pdev->width - (dev_l_margin(pdev) + dev_r_margin(pdev)) * pdev->x_pixels_per_inch;
-          int bm = pdev->height - (dev_t_margin(pdev) + dev_b_margin(pdev)) * pdev->y_pixels_per_inch;
+          int rm = (int)(pdev->width - (dev_l_margin(pdev) + dev_r_margin(pdev)) * pdev->x_pixels_per_inch);
+          int bm = (int)(pdev->height - (dev_t_margin(pdev) + dev_b_margin(pdev)) * pdev->y_pixels_per_inch);
           /* Draw black rectangle */
           gp_fprintf(prn_stream,
                      "%c{%c%da%c%de%c;;;3}",
@@ -941,8 +941,8 @@ lips_job_start(gx_device_printer * pdev, lips_printer_type ptype, gp_file * prn_
     gx_device_lips *const lips = (gx_device_lips *) pdev;
     gx_device_lips4 *const lips4 = (gx_device_lips4 *) pdev;
     int prev_paper_size, prev_paper_width, prev_paper_height, paper_size;
-    int width = pdev->MediaSize[0];
-    int height = pdev->MediaSize[1];
+    int width = (int)pdev->MediaSize[0];
+    int height = (int)pdev->MediaSize[1];
     int tm, lm, rm, bm;
 
     if (pdev->PageCount == 0) {
@@ -1168,14 +1168,14 @@ lips_job_start(gx_device_printer * pdev, lips_printer_type ptype, gp_file * prn_
     }
     if (prev_paper_size != paper_size) {
         /* Top Margin: 63/300 inch + 5 mm */
-        tm = (63. / 300. + 5. / MMETER_PER_INCH - dev_t_margin(pdev)) * pdev->x_pixels_per_inch;
+        tm = (int)((63. / 300. + 5. / MMETER_PER_INCH - dev_t_margin(pdev)) * pdev->x_pixels_per_inch);
         if (tm > 0)
             gp_fprintf(prn_stream, "%c%dk", LIPS_CSI, tm);
         if (tm < 0)
             gp_fprintf(prn_stream, "%c%de", LIPS_CSI, -tm);
 
         /* Left Margin: 5 mm left */
-        lm = (5. / MMETER_PER_INCH - dev_l_margin(pdev)) * pdev->x_pixels_per_inch;
+        lm = (int)((5. / MMETER_PER_INCH - dev_l_margin(pdev)) * pdev->x_pixels_per_inch);
         if (lm > 0)
             gp_fprintf(prn_stream, "%c%dj", LIPS_CSI, lm);
         if (lm < 0)
@@ -1185,10 +1185,10 @@ lips_job_start(gx_device_printer * pdev, lips_printer_type ptype, gp_file * prn_
         gp_fprintf(prn_stream, "%c0;2t", LIPS_CSI);
 
         /* Bottom Margin: height */
-        bm = pdev->height - (dev_t_margin(pdev) + dev_b_margin(pdev)) * pdev->y_pixels_per_inch;
+        bm = (int)(pdev->height - (dev_t_margin(pdev) + dev_b_margin(pdev)) * pdev->y_pixels_per_inch);
         gp_fprintf(prn_stream, "%c%de", LIPS_CSI, bm);
         /* Right Margin: width */
-        rm = pdev->width - (dev_l_margin(pdev) + dev_r_margin(pdev)) * pdev->x_pixels_per_inch;
+        rm = (int)(pdev->width - (dev_l_margin(pdev) + dev_r_margin(pdev)) * pdev->x_pixels_per_inch);
         gp_fprintf(prn_stream, "%c%da", LIPS_CSI, rm);
         gp_fprintf(prn_stream, "%c1;3t", LIPS_CSI);
